+++ packages/react-charts/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.4](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-charts@1.4.4-beta.0...@tecsinapse/react-charts@1.4.4) (2024-06-10)
+
+**Note:** Version bump only for package @tecsinapse/react-charts
+
+## [1.4.4-beta.0](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-charts@1.4.1-beta.0...@tecsinapse/react-charts@1.4.4-beta.0) (2024-06-10)
+
+**Note:** Version bump only for package @tecsinapse/react-charts
+
 ## [1.4.3](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-charts@1.4.2...@tecsinapse/react-charts@1.4.3) (2024-05-23)
 
 **Note:** Version bump only for package @tecsinapse/react-charts